move Screens package outside of the navigation package
